atlas-v2/requirements.txt
# Core dependencies
python-telegram-bot>=20.7
openai>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0

# Memory system
numpy>=1.24.0

# Browser automation (optional)
playwright>=1.40.0

# Sentence embeddings (optional, for semantic search)
# sentence-transformers>=2.2.0

# Database
# sqlite-vec  # For vector search extension (optional)
